Moved System.Drawing dependent methods in ImageBox to extensions class. Issue #83
